All scripts: fix undefined lock id & log file name
Some scripts compute the log file name and lock id from the arguments of
the script.
The log file name is required for logging information and errors.
Until now the log functions were sometimes called before the log file
name was known.

The script were corrected so that:
- the log file name and lock-id are not called anymore before the end of
the fct parseInputParams()
- error detected in the fct parseInputParams() are only echoed
- the echoed errors are piped to the log file after the execution of the
fct parseInputParams()