#34 - `backupData.sh`: New meaning of `fsDest` argument (#38)
Fixes #34

ATTENTION: This change BREAKS the compatibility with the previous
version of the command line interface.

backupData.sh: Change the meaning of the command line argument `fsDest`
(#34):

* Before the change the backup was done in `tank_backup/tank` if
`fsSource=tank` and `fsDest=tank_backup`.
* After he change the backup is done in `tank_backup` directly.

Rationale: This gives more control over the destination FS.
This is required to support restoring a backup using `backupData.sh`
(only source and destination filesystems must be slipped in the call)